+++ sdk/resourcemanager/netapp/armnetapp/models.go
@@ -90,6 +90,9 @@ type AccountProperties struct {
 	// Encryption settings
 	Encryption *AccountEncryption
 
+	// LDAP Configuration for the account.
+	LdapConfiguration *LdapConfiguration
+
 	// Domain for NFSv4 user ID mapping. This property will be set for all NetApp accounts in the subscription and region and
 	// only affect non ldap NFSv4 volumes.
 	NfsV4IDDomain *string
@@ -370,7 +373,9 @@ type BackupStatus struct {
 	// READ-ONLY; Displays the last transfer type
 	LastTransferType *string
 
-	// READ-ONLY; The status of the backup
+	// READ-ONLY; The mirror state property describes the current status of data replication for a backup. It provides insight
+	// into whether the data is actively being mirrored, if the replication process has been
+	// paused, or if it has yet to be initialized.
 	MirrorState *MirrorState
 
 	// READ-ONLY; Status of the backup mirror relationship
@@ -462,6 +467,146 @@ type BreakReplicationRequest struct {
 	ForceBreakReplication *bool
 }
 
+// Bucket resource
+type Bucket struct {
+	// Bucket properties
+	Properties *BucketProperties
+
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
+	ID *string
+
+	// READ-ONLY; The name of the resource
+	Name *string
+
+	// READ-ONLY; Azure Resource Manager metadata containing createdBy and modifiedBy information.
+	SystemData *SystemData
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string
+}
+
+// BucketCredentialsExpiry - The bucket's Access and Secret key pair Expiry Time expressed as the number of days from now.
+type BucketCredentialsExpiry struct {
+	// The number of days from now until the newly generated Access and Secret key pair will expire.
+	KeyPairExpiryDays *int32
+}
+
+// BucketGenerateCredentials - Bucket Access Key, Secret Key, and Expiry date and time of the key pair
+type BucketGenerateCredentials struct {
+	// READ-ONLY; The Access Key that is required along with the Secret Key to access the bucket.
+	AccessKey *string
+
+	// READ-ONLY; The bucket's Access and Secret key pair expiry date and time (in UTC).
+	KeyPairExpiry *time.Time
+
+	// READ-ONLY; The Secret Key that is required along with the Access Key to access the bucket.
+	SecretKey *string
+}
+
+// BucketList - List of volume bucket resources
+type BucketList struct {
+	// URL to get the next set of results.
+	NextLink *string
+
+	// List of volume buckets
+	Value []*Bucket
+}
+
+// BucketPatch - Bucket resource
+type BucketPatch struct {
+	// Bucket properties
+	Properties *BucketPatchProperties
+
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
+	ID *string
+
+	// READ-ONLY; The name of the resource
+	Name *string
+
+	// READ-ONLY; Azure Resource Manager metadata containing createdBy and modifiedBy information.
+	SystemData *SystemData
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string
+}
+
+// BucketPatchProperties - Bucket resource properties for a Patch operation
+type BucketPatchProperties struct {
+	// File System user having access to volume data. For Unix, this is the user's uid and gid. For Windows, this is the user's
+	// username. Note that the Unix and Windows user details are mutually exclusive,
+	// meaning one or other must be supplied, but not both.
+	FileSystemUser *FileSystemUser
+
+	// The volume path mounted inside the bucket.
+	Path *string
+
+	// Access permissions for the bucket. Either ReadOnly or ReadWrite.
+	Permissions *BucketPatchPermissions
+
+	// Properties of the server managing the lifecycle of volume buckets
+	Server *BucketServerPatchProperties
+
+	// READ-ONLY; Provisioning state of the resource
+	ProvisioningState *NetAppProvisioningState
+}
+
+// BucketProperties - Bucket resource properties
+type BucketProperties struct {
+	// File System user having access to volume data. For Unix, this is the user's uid and gid. For Windows, this is the user's
+	// username. Note that the Unix and Windows user details are mutually exclusive,
+	// meaning one or other must be supplied, but not both.
+	FileSystemUser *FileSystemUser
+
+	// The volume path mounted inside the bucket. The default is the root path '/' if no value is provided when the bucket is
+	// created.
+	Path *string
+
+	// Access permissions for the bucket. Either ReadOnly or ReadWrite. The default is ReadOnly if no value is provided during
+	// bucket creation.
+	Permissions *BucketPermissions
+
+	// Properties of the server managing the lifecycle of volume buckets
+	Server *BucketServerProperties
+
+	// READ-ONLY; Provisioning state of the resource
+	ProvisioningState *NetAppProvisioningState
+
+	// READ-ONLY; The bucket credentials status. There states:
+	// "NoCredentialsSet": Access and Secret key pair have not been generated. "CredentialsExpired": Access and Secret key pair
+	// have expired. "Active": The certificate has been installed and credentials are
+	// unexpired.
+	Status *CredentialsStatus
+}
+
+// BucketServerPatchProperties - Properties of the server managing the lifecycle of volume buckets
+type BucketServerPatchProperties struct {
+	// A base64-encoded PEM file, which includes both the bucket server's certificate and private key. It is used to authenticate
+	// the user and allows access to volume data in a read-only manner.
+	CertificateObject *string
+
+	// The host part of the bucket URL, resolving to the bucket IP address and allowed by the server certificate.
+	Fqdn *string
+}
+
+// BucketServerProperties - Properties of the server managing the lifecycle of volume buckets
+type BucketServerProperties struct {
+	// A base64-encoded PEM file, which includes both the bucket server's certificate and private key. It is used to authenticate
+	// the user and allows access to volume data in a read-only manner.
+	CertificateObject *string
+
+	// The host part of the bucket URL, resolving to the bucket IP address and allowed by the server certificate.
+	Fqdn *string
+
+	// READ-ONLY; Certificate Common Name taken from the certificate installed on the bucket server
+	CertificateCommonName *string
+
+	// READ-ONLY; The bucket server's certificate expiry date.
+	CertificateExpiryDate *time.Time
+
+	// READ-ONLY; The bucket server's IPv4 address
+	IPAddress *string
+}
+
 // CapacityPool - Capacity pool resource
 type CapacityPool struct {
 	// REQUIRED; The geo-location where the resource lives
@@ -550,6 +695,12 @@ type CheckAvailabilityResponse struct {
 	Reason *InAvailabilityReasonType
 }
 
+// CifsUser - The effective CIFS username when accessing the volume data.
+type CifsUser struct {
+	// The CIFS user's username
+	Username *string
+}
+
 // ClusterPeerCommandResponse - Information about cluster peering process
 type ClusterPeerCommandResponse struct {
 	// A command that needs to be run on the external ONTAP to accept cluster peering. Will only be present if clusterPeeringStatus
@@ -598,7 +749,7 @@ type Dimension struct {
 
 // EncryptionIdentity - Identity used to authenticate with key vault.
 type EncryptionIdentity struct {
-	// ClientId of the multi-tenant AAD Application. Used to access cross-tenant keyvaults.
+	// ClientId of the multi-tenant AAD Application. Used to access cross-tenant KeyVaults.
 	FederatedClientID *string
 
 	// The ARM resource identifier of the user assigned identity used to authenticate with key vault. Applicable if identity.type
@@ -681,6 +832,17 @@ type FilePathAvailabilityRequest struct {
 	AvailabilityZone *string
 }
 
+// FileSystemUser - File System user having access to volume data. For Unix, this is the user's uid and gid. For Windows,
+// this is the user's username. Note that the Unix and Windows user details are mutually exclusive,
+// meaning one or other must be supplied, but not both.
+type FileSystemUser struct {
+	// The effective CIFS username when accessing the volume data.
+	CifsUser *CifsUser
+
+	// The effective NFS User ID and Group ID when accessing the volume data.
+	NfsUser *NfsUser
+}
+
 // GetGroupIDListForLDAPUserRequest - Get group Id list for LDAP User request
 type GetGroupIDListForLDAPUserRequest struct {
 	// REQUIRED; username is required to fetch the group to which user is part of
@@ -755,6 +917,24 @@ type KeyVaultProperties struct {
 	Status *KeyVaultStatus
 }
 
+// LdapConfiguration - LDAP configuration
+type LdapConfiguration struct {
+	// The CN host name used while generating the certificate, LDAP Over TLS requires the CN host name to create DNS host entry.
+	CertificateCNHost *string
+
+	// Name of the LDAP configuration domain
+	Domain *string
+
+	// Specifies whether or not the LDAP traffic needs to be secured via TLS.
+	LdapOverTLS *bool
+
+	// List of LDAP server IP addresses (IPv4 only) for the LDAP domain.
+	LdapServers []*string
+
+	// When LDAP over SSL/TLS is enabled, the LDAP client is required to have base64 encoded ldap servers CA certificate.
+	ServerCACertificate *string
+}
+
 // LdapSearchScopeOpt - LDAP search scope
 type LdapSearchScopeOpt struct {
 	// This specifies the group DN, which overrides the base DN for group lookups.
@@ -765,6 +945,12 @@ type LdapSearchScopeOpt struct {
 
 	// This specifies the user DN, which overrides the base DN for user lookups.
 	UserDN *string
+}
+
+// ListQuotaReportResponse - Quota Report for volume
+type ListQuotaReportResponse struct {
+	// List of volume quota report records
+	Value []*QuotaReport
 }
 
 // ListReplications - List Replications
@@ -906,6 +1092,15 @@ type NetworkSiblingSet struct {
 
 	// READ-ONLY; Gets the status of the NetworkSiblingSet at the time the operation was called.
 	ProvisioningState *NetworkSiblingSetProvisioningState
+}
+
+// NfsUser - The effective NFS User ID and Group ID when accessing the volume data.
+type NfsUser struct {
+	// The NFS user's GID
+	GroupID *int64
+
+	// The NFS user's UID
+	UserID *int64
 }
 
 // NicInfo - NIC information and list of volumes for which the NIC has the primary mount ip address.
@@ -1057,6 +1252,67 @@ type QuotaAvailabilityRequest struct {
 	Type *CheckQuotaNameResourceTypes
 }
 
+// QuotaItem - Information regarding Quota Item.
+type QuotaItem struct {
+	// QuotaItem properties
+	Properties *QuotaItemProperties
+
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
+	ID *string
+
+	// READ-ONLY; The name of the resource
+	Name *string
+
+	// READ-ONLY; Azure Resource Manager metadata containing createdBy and modifiedBy information.
+	SystemData *SystemData
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string
+}
+
+// QuotaItemList - List of Quota Items
+type QuotaItemList struct {
+	// URL to get the next set of results.
+	NextLink *string
+
+	// A list of QuotaItems
+	Value []*QuotaItem
+}
+
+// QuotaItemProperties - QuotaItem Properties
+type QuotaItemProperties struct {
+	// READ-ONLY; The current quota value.
+	Current *int32
+
+	// READ-ONLY; The default quota value.
+	Default *int32
+
+	// READ-ONLY; The usage quota value.
+	Usage *int32
+}
+
+// QuotaReport - Quota report record properties
+type QuotaReport struct {
+	// Flag to indicate whether the quota is derived from default quota.
+	IsDerivedQuota *bool
+
+	// Percentage of used size compared to total size.
+	PercentageUsed *float32
+
+	// Specifies the total size limit in kibibytes for the user/group quota.
+	QuotaLimitTotalInKiBs *int64
+
+	// Specifies the current usage in kibibytes for the user/group quota.
+	QuotaLimitUsedInKiBs *int64
+
+	// UserID/GroupID/SID based on the quota target type. UserID and groupID can be found by running ‘id’ or ‘getent’ command
+	// for the user or group and SID can be found by running
+	QuotaTarget *string
+
+	// Type of quota
+	QuotaType *Type
+}
+
 // ReestablishReplicationRequest - Re-establish request object supplied in the body of the operation.
 type ReestablishReplicationRequest struct {
 	// Resource id of the source volume for the replication
@@ -1163,6 +1419,21 @@ type ReplicationObject struct {
 	// READ-ONLY; Indicates whether the local volume is the source or destination for the Volume Replication
 	EndpointType *EndpointType
 
+	// READ-ONLY; Contains human-readable instructions on what the next step is to finish the external replication setup.
+	ExternalReplicationSetupInfo *string
+
+	// READ-ONLY; Property that only applies to external replications. Provides a machine-readable value for the status of the
+	// external replication setup.
+	ExternalReplicationSetupStatus *ExternalReplicationSetupStatus
+
+	// READ-ONLY; The mirror state property describes the current status of data replication for a replication. It provides insight
+	// into whether the data is actively being mirrored, if the replication process has been
+	// paused, or if it has yet to be initialized.
+	MirrorState *MirrorState
+
+	// READ-ONLY; The status of the Volume Replication
+	RelationshipStatus *RelationshipStatus
+
 	// READ-ONLY; Id
 	ReplicationID *string
 }
@@ -1175,14 +1446,16 @@ type ReplicationStatus struct {
 	// Replication health check
 	Healthy *bool
 
-	// The status of the replication
-	MirrorState *MirrorState
-
 	// Status of the mirror relationship
 	RelationshipStatus *RelationshipStatus
 
 	// The progress of the replication
 	TotalProgress *string
+
+	// READ-ONLY; The mirror state property describes the current status of data replication for a replication. It provides insight
+	// into whether the data is actively being mirrored, if the replication process has been
+	// paused, or if it has yet to be initialized.
+	MirrorState *MirrorState
 }
 
 // ResourceNameAvailabilityRequest - Resource name availability request content.
@@ -1205,7 +1478,9 @@ type RestoreStatus struct {
 	// READ-ONLY; Restore health status
 	Healthy *bool
 
-	// READ-ONLY; The status of the restore
+	// READ-ONLY; The mirror state property describes the current status of data replication for a restore. It provides insight
+	// into whether the data is actively being mirrored, if the replication process has been
+	// paused, or if it has yet to be initialized.
 	MirrorState *MirrorState
 
 	// READ-ONLY; Status of the restore SnapMirror relationship
@@ -1354,42 +1629,6 @@ type SnapshotRestoreFiles struct {
 type SnapshotsList struct {
 	// A list of Snapshots
 	Value []*Snapshot
-}
-
-// SubscriptionQuotaItem - Information regarding Subscription Quota Item.
-type SubscriptionQuotaItem struct {
-	// SubscriptionQuotaItem properties
-	Properties *SubscriptionQuotaItemProperties
-
-	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
-	ID *string
-
-	// READ-ONLY; The name of the resource
-	Name *string
-
-	// READ-ONLY; Azure Resource Manager metadata containing createdBy and modifiedBy information.
-	SystemData *SystemData
-
-	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
-	Type *string
-}
-
-// SubscriptionQuotaItemList - List of Subscription Quota Items
-type SubscriptionQuotaItemList struct {
-	// URL to get the next set of results.
-	NextLink *string
-
-	// A list of SubscriptionQuotaItems
-	Value []*SubscriptionQuotaItem
-}
-
-// SubscriptionQuotaItemProperties - SubscriptionQuotaItem Properties
-type SubscriptionQuotaItemProperties struct {
-	// READ-ONLY; The current quota value.
-	Current *int32
-
-	// READ-ONLY; The default quota value.
-	Default *int32
 }
 
 // SubvolumeInfo - Subvolume Information properties
@@ -1950,8 +2189,14 @@ type VolumeProperties struct {
 	// = 'Microsoft.KeyVault'.
 	KeyVaultPrivateEndpointResourceID *string
 
+	// Language supported for volume.
+	Language *VolumeLanguage
+
 	// Specifies whether LDAP is enabled or not for a given NFS volume.
 	LdapEnabled *bool
+
+	// Specifies the type of LDAP server for a given NFS volume.
+	LdapServerType *LdapServerType
 
 	// The original value of the network features type available to the volume at the time it was created.
 	NetworkFeatures *NetworkFeatures
@@ -1999,14 +2244,18 @@ type VolumeProperties struct {
 	// (2) and sticky (1) attributes. Second digit selects permission for the owner of
 	// the file: read (4), write (2) and execute (1). Third selects permissions for other users in the same group. the fourth
 	// for other users not in the group. 0755 - gives read/write/execute permissions to
-	// owner and read/execute to group and other users.
+	// owner and read/execute to group and other users. Avoid passing null value for unixPermissions in volume update operation,
+	// As per the behavior, If Null value is passed then user-visible unixPermissions
+	// value will became null, and user will not be able to get unixPermissions value. On safer side, actual unixPermissions value
+	// on volume will remain as its last saved value only.
 	UnixPermissions *string
 
 	// Volume spec name is the application specific designation or identifier for the particular volume in a volume group for
 	// e.g. data, log
 	VolumeSpecName *string
 
-	// What type of volume is this. For destination volumes in Cross Region Replication, set type to DataProtection
+	// What type of volume is this. For destination volumes in Cross Region Replication, set type to DataProtection. For creating
+	// clone volume, set type to ShortTermClone
 	VolumeType *string
 
 	// READ-ONLY; Actual throughput in MiB/s for auto qosType volumes calculated based on size and serviceLevel
@@ -2144,8 +2393,8 @@ type VolumeQuotaRulesProperties struct {
 	// Type of quota
 	QuotaType *Type
 
-	// READ-ONLY; Gets the status of the VolumeQuotaRule at the time the operation was called.
-	ProvisioningState *ProvisioningState
+	// READ-ONLY; Provisioning state of the resource
+	ProvisioningState *NetAppProvisioningState
 }
 
 // VolumeRelocationProperties - Volume relocation properties